# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Volcanion Stress Test Tool API
  description: |
    A powerful HTTP load testing and stress testing tool built in Go.
    
    ## Authentication
    
    The API supports two authentication methods:
    
    1. **JWT Bearer Token**: Obtain a token via `/api/v1/auth/login` and include it in the `Authorization` header.
    2. **API Key**: Include your API key in the `X-API-Key` header.
    
    ## Rate Limiting
    
    API endpoints are rate-limited to 100 requests per minute per IP address.
    When rate limited, you'll receive a `429 Too Many Requests` response.
    
    ## Versioning
    
    All API endpoints are versioned under `/api/v1/`. The current version is returned 
    in the `X-API-Version` response header.
    
  version: 1.0.0
  contact:
    name: Volcanion Team
    url: https://github.com/volcanion-company/volcanion-stress-test-tool
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.volcanion.io
    description: Production server (example)
...
paths:
...
  /api/v1/auth/api-keys:
    post:
      summary: Create API Key
      description: Create a new API key for programmatic access
      operationId: createApiKey
      tags:
        - Authentication
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPIKeyRequest'
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAPIKeyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      summary: List API Keys
      description: List all API keys for the authenticated user
      operationId: listApiKeys
      tags:
        - Authentication
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIKey'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/api-keys/{id}:
    delete:
      summary: Revoke API Key
      description: Revoke an existing API key
      operationId: revokeApiKey
      tags:
        - Authentication
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: API key revoked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/audit/logs:
    get:
      summary: Get Audit Logs
      description: Get audit logs (admin only)
      operationId: getAuditLogs
      tags:
        - Audit
      security:
        - bearerAuth: []
      parameters:
        - name: start_time
          in: query
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          schema:
            type: string
            format: date-time
        - name: action
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
...
    CreateAPIKeyRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Descriptive name for the API key
        role:
          type: string
          enum: [admin, user, readonly]
          default: user
        expires_at:
          type: string
          format: date-time
          description: Optional expiration date
    CreateAPIKeyResponse:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
          description: The API key (only shown once)
        name:
          type: string
        role:
          type: string
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    APIKey:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
        last_used_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
    AuditLog:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        user_id:
          type: string
        action:
          type: string
        resource:
          type: string
        resource_id:
          type: string
        details:
          type: object
        ip_address:
          type: string